+++ client/src/utils/titleCase.ts
